Module glenside::codegen

Functions

c_allocation_string
Creates a representation (currently just a string with a C declaration) of an allocation, given the name, shape, dtype, and any prefix.

c_array_string
Gives the signature of a C array given the datatype, name, and shape. Useful for declaring arrays and array-type function arguments.

c_assignment_string
Creates a representation (currently just a string with a C definition) of an allocation and an assignment.

codegen
Returns c code.

create_hardware_design_monolithic
Create a monolithic systolic array hardware design from an expression, where we use just one systolic array. (row, col): the rows/cols of the systolic array.

create_hardware_design_no_sharing
Create a hardware design from an expression, creating a unique atom for each unique node (eclass or id).

find_vars
Finds all symbols in a program, and return their names.

generate_worklist_for_codegen
Generates a worklist for codegen recursively, given an egraph and an eclass id to start from.