Function glenside::codegen::create_hardware_design_monolithic

pub fn create_hardware_design_monolithic(
    expr: &EGraph<Language, MyAnalysis>, 
    (row, col): (usize, usize)
) -> (HashMap<Id, usize>, Vec<Atom>)

Create a monolithic systolic array hardware design from an expression, where we use just one systolic array. (row, col): the rows/cols of the systolic array.

use glenside::hw_design_language::*;
use glenside::language::*;
use glenside::codegen::*;
use egg::*;
use std::collections::HashMap;
use std::str::FromStr;

let expr = RecExpr::from_str(
    "
    (systolic-array 32 32
     (access
      (systolic-array 32 32
       (access (access-tensor t0) 1)
       (access (access-tensor t1) 0)
      )
      1
     )
     (access
      (systolic-array 32 32
       (access (access-tensor t2) 1)
       (access (access-tensor t3) 0)
      )
      0
     )
    )",
)
.unwrap();

let mut map = HashMap::default();
map.insert("t0".to_string(), vec![32, 32]);
map.insert("t1".to_string(), vec![32, 32]);
map.insert("t2".to_string(), vec![32, 32]);
map.insert("t3".to_string(), vec![32, 32]);

let mut egraph = EGraph::new(MyAnalysis { name_to_shape: map });
egraph.add_expr(&expr);

let (hw_map, hw_design) = create_hardware_design_monolithic(&egraph, (32, 32));

assert_eq!(hw_design.len(), 1);
match &hw_design[0] {
    Atom {
        name,
        id: 0,
        config:
            AtomConfig::SystolicArrayWeightStationary(SystolicArrayWeightStationaryParams {
                dtype: DType::Fp32,
                rows: 32,
                cols: 32,
            }),
    } => assert_eq!(*name, "systolic_array_0".to_string()),
    _ => panic!(),
};

assert_eq!(hw_map.len(), 3);
for (_id, val) in hw_map.iter() {
    assert_eq!(*val, 0);
}