Module glenside::extraction

Structs

MonolithicCostFunction
A cost function to extract a design using a single size of systolic array.

SimpleCostFunction
This cost function applies the bare minimum amount of logic to produce a valid hardware/software program. Most importantly, it blocks Compute nodes from being extracted, as these nodes should be replaced by hardware atoms (or, perhaps in the future, kernel calls). It also filters out old Glenside constructs.

Functions

find_all_systolic_array_configurations