Module glenside::extraction::ilp

Integer Linear Programming-based egraph extractor for Glenside

Neural networks expressed in Glenside are so large that the standard egg::Extractor cannot be used; the egg::CostFunctions end up overflowing [usize::Max]! This is a problem of common subexpressions appearing many, many times. For example, when access-slice operators slice the same tensor multiple times, they replicate the sliced tensor’s subexpression each time, blowing up the size of the overall expression significantly.

ILP extraction (and other complex extraction methods) have generally been the answer for these types of extraction problems in egg.

Our implementation will draw heavily from Remy Wang’s SPORES paper and warp repository.

Structs

EGraphLpProblem

Thin wrapper over [lp_modeler::LpProblem].

Functions

create_generic_egraph_lp_model

From an egraph, create an LP model with a few useful base constraints

extract_single_expression

Returns an [EGraph] with just one enode per eclass, according to the results of the ILP problem. I could use a [RecExpr] here, but all of the post-processing we do (i.e. codegen) prefers to have an [EGraph] with our analysis attached. Also returns a map, mapping the Ids in the old egraph to their Ids in the new expression.

filter_by_enode_type

filter_obviously_less_preferable_nodes

Filtering function that drops enodes if this eclass contains obviously-extractable constructs.

filter_useless_access_flattens

Filtering function which filters out nodes which form simple loops. Returns false if and only if this node is an access-flatten which is in an eclass by itself, and if the eclass it points to has an access-reshape node that points right back to this node’s eclass.

filter_useless_pad_slice_loops

Filtering function which filters out useless pad/slice loops