Function glenside::extraction::ilp::create_generic_egraph_lp_model

pub fn create_generic_egraph_lp_model<'a>(
    env: &'a Env, 
    egraph: &'a EGraph<Language, MyAnalysis>, 
    filter_enode: impl Fn(&Language, Id, &EGraph<Language, MyAnalysis>) -> bool, 
    roots: &[Id], 
    name: &'static str
) -> EGraphLpProblem<'a>

From an egraph, create an LP model with a few useful base constraints

We create three types of variables: eclass variables, enode variables, and topological sorting variables. See the definition of EGraphLpProblem for more detail on these variables.

The filter_enode argument is a function which, given a reference to the enode, plus the id of its containing [EClass] and a reference to the egg::EGraph itself, determines whether to create a variable in the ILP problem for this enode. This can be used to filter “useless” enodes.

Code taken from Remy Wang’s warp repository.