Function glenside::extraction::ilp::filter_useless_access_flattens

pub fn filter_useless_access_flattens(
    enode: &Language, 
    eclass_id: Id, 
    egraph: &EGraph<Language, MyAnalysis>
) -> bool

Filtering function which filters out nodes which form simple loops. Returns false if and only if this node is an access-flatten which is in an eclass by itself, and if the eclass it points to has an access-reshape node that points right back to this node’s eclass.