Module glenside::language::from_relay

Functions

access
Given an access and axis, add access expression accessing access at axis

access_concatenate
Concatenate accesses

access_insert_axis
Given an access and axis, add access expression accessing access at axis

access_pad
Pad an access

access_pair
Given input accesses and an axis, re-accesses and pairs both accesses

access_shape
Create access shape literal

access_slice
Slice an access

access_transpose

compute
Given the input access and compute type, add compute expression

conv2d

from_relay
Convert Relay IRModule to Glenside RecExpr.

list

shape
Create a shape

shape_from_type
Get shape from type