Function glenside::language::from_relay::from_relay

pub fn from_relay(
    module: &IRModule, 
    simplify_batch_norm_for_inference_hack: bool, 
    use_opaque_operators_for: &Vec<RelayOperator>
) -> (RecExpr<Language>, Vec<(String, Vec<usize>)>)

Convert Relay IRModule to Glenside RecExpr.

Returns the RecExpr, along with a Vec mapping symbols to their shapes. Note that the shapes are a Vec rather than a HashMap to preserve ordering.

If simplify_batch_norm_for_inference_hack is enabled, any batch norm found in the Relay program will be converted to a RelayOperatorCall node in Glenside. Otherwise, the function will panic.

TODO(@gussmith23) Get rid of this hack eventually

use_opaque_operators_for is a list of RelayOperators which should be replaced with opaque versions of their operators, instead of expanding them into Glenside code. This is a hack; this allows us to mostly ignore operators that we don’t yet do rewrites over (e.g. batch norms, pooling, softmax) and allows us to more easily generate somewhat good code for them.