Function glenside::language::interpreter::interpret_from_str

pub fn interpret_from_str<DataType: 'static>(
    program: &str, 
    env: &Environment<'_, DataType>
) -> Value<DataType> where
    DataType: Copy + Mul<Output = DataType> + Div<Output = DataType> + Neg<Output = DataType> + Sum + One + Zero + PartialOrd + Bounded + Exp + Sqrt + FromNotNanFloat64Literal + ScalarOperand,
    usize: AsPrimitive<DataType>, 

Simple wrapper over interpret.

This was created for the web demo. Specifically, this lets us avoid having the web demo depend on egg directly, thus avoiding versioning mismatches between glenside’s egg and the web demo’s egg. There may be a much better way to handle that, though.

use glenside::language::interpreter::interpret_from_str;
use glenside::language::interpreter::Value;
use std::collections::HashMap;
use ndarray::Dimension;

match interpret_from_str::<i64>("(access-shape (shape 1 2) (shape 3 4))", &HashMap::default()) {
    Value::AccessShape(shape, access_axis) => {
        assert_eq!(shape.slice(), &[1, 2, 3, 4]);
        assert_eq!(access_axis, 2);
    }
    _ => panic!(),
}