Function glenside::language::interpreter::interpret

pub fn interpret<DataType: 'static>(
    expr: &RecExpr<Language>, 
    index: usize, 
    env: &Environment<'_, DataType>
) -> Value<DataType> where
    DataType: Copy + Mul<Output = DataType> + Div<Output = DataType> + Neg<Output = DataType> + Sum + One + Zero + PartialOrd + Bounded + Exp + Sqrt + FromNotNanFloat64Literal + ScalarOperand,
    usize: AsPrimitive<DataType>, 

Interpret a Glenside expression

Generally, DataType can be inferred from the environment passed in. If your expression doesn’t actually use any tensor values, and the environment is empty, you can choose an arbitrary type: e.g. interpret::<i64>(...):

use egg::RecExpr;
use glenside::language::Language;
use glenside::language::interpreter::interpret;
use glenside::language::interpreter::Value;
use std::str::FromStr;
use std::collections::HashMap;
use ndarray::Dimension;

let expr = RecExpr::<Language>::from_str("(access-shape (shape 1 2) (shape 3 4))").unwrap();
match interpret::<i64>(&expr, expr.as_ref().len() - 1, &HashMap::default()) {
    Value::AccessShape(shape, access_axis) => {
        assert_eq!(shape.slice(), &[1, 2, 3, 4]);
        assert_eq!(access_axis, 2);
    }
    _ => panic!(),
}