Trait glenside::language::RangeSet

pub trait RangeSet {
    type Index;
    fn insert_elements(
        &mut self, 
        strategy: RangeInsertStrategy, 
        index: Self::Index, 
        num_elements_inserted: usize
    );

    fn remove_elements(
        &mut self, 
        index: Self::Index, 
        num_elements_removed: usize
    );

    fn covered(&self, range: (Self::Index, Self::Index)) -> bool;

    fn add_range(&mut self, range: (Self::Index, Self::Index));
}

Associated Types

type Index

Loading content...
Required methods

fn insert_elements(
&mut self,
strategy: RangeInsertStrategy,
index: Self::Index,
num_elements_inserted: usize
)

Updates ranges as if num_elements_inserted elements were inserted at index, according to strategy.

fn remove_elements(&mut self, index: Self::Index, num_elements_removed: usize)

Updates ranges as if num_elements_removed elements were removed at index.

fn covered(&self, range: (Self::Index, Self::Index)) -> bool

Checks whether range is covered by the ranges in this set.

fn add_range(&mut self, range: (Self::Index, Self::Index))

Adds range. Ranges are half-open.

Loading content...